# Conflicts:
#	Classes/WDArrowheadCell.m
#	Classes/WDBlendModeController.m
#	Classes/WDCanvas.h
#	Classes/WDCanvas.m
#	Classes/WDColorWell.m
#	Classes/WDFillController.m
#	Classes/WDFontController.m
#	Classes/WDGradientController.m
#	Classes/WDHelpController.m
#	Classes/WDHueSaturationController.m
#	Classes/WDImportController.m
#	Classes/WDLayerController.m
#	Classes/WDMenuItem.h
#	Classes/WDMenuItem.m
#	Classes/WDPageSizeController.m
#	Classes/WDSamplesController.m
#	Classes/WDSelectionView.m
#	Classes/WDSettingsController.m
#	Classes/WDShadowController.m
#	Classes/WDShadowWell.m
#	Classes/WDStrokeController.m
#	Classes/WDSwatchController.m
#	Classes/WDTextController.m
#	Classes/WDThumbnailView.m
#	Classes/WDUnitsController.m
#	Inkpad-Core/Model/WDAbstractPath.m
#	Inkpad-Core/Model/WDBezierNode.m
#	Inkpad-Core/Model/WDColor.m
#	Inkpad-Core/Model/WDCompoundPath.m
#	Inkpad-Core/Model/WDDrawing.m
#	Inkpad-Core/Model/WDFillTransform.m
#	Inkpad-Core/Model/WDGradient.m
#	Inkpad-Core/Model/WDGradientStop.m
#	Inkpad-Core/Model/WDGroup.m
#	Inkpad-Core/Model/WDImage.m
#	Inkpad-Core/Model/WDImageData.m
#	Inkpad-Core/Model/WDLayer.m
#	Inkpad-Core/Model/WDPath.m
#	Inkpad-Core/Model/WDShadow.m
#	Inkpad-Core/Model/WDStrokeStyle.m
#	Inkpad-Core/Model/WDStylable.m
#	Inkpad-Core/Model/WDText.m
#	Inkpad-Core/Model/WDTextPath.m
#	Openclipart/OCAViewController.m

+++ Inkpad-Core/Model/WDDrawing.h
@@ -10,8 +10,7 @@
 //
 
 #import <UIKit/UIKit.h>
-#import <CoreData/CoreData.h>
-#import <Foundation/Foundation.h>
+
 #import "WDDocumentProtocol.h"
 #import "WDStrokeStyle.h"
 
@@ -24,8 +23,8 @@
 @class WDPickResult;
 @class WDRulerUnit;
 
-extern const float kMinimumDrawingDimension;
-extern const float kMaximumDrawingDimension;
+extern const CGFloat kMinimumDrawingDimension;
+extern const CGFloat kMaximumDrawingDimension;
 
 enum
 {
@@ -41,7 +40,7 @@ typedef struct
 	UInt32  flags;
 } WDRenderingMetaData;
 
-WDRenderingMetaData WDRenderingMetaDataMake(float scale, UInt32 flags);
+WDRenderingMetaData WDRenderingMetaDataMake(CGFloat scale, UInt32 flags);
 BOOL WDRenderingMetaDataOutlineOnly(WDRenderingMetaData metaData);
 
 @protocol WDDocumentProtocol;
@@ -61,8 +60,8 @@ BOOL WDRenderingMetaDataOutlineOnly(WDRenderingMetaData metaData);
 @property (nonatomic, strong) NSUndoManager* undoManager;
 @property (nonatomic, weak) id<WDDocumentProtocol> document;
 
-@property (nonatomic, assign) float width;
-@property (nonatomic, assign) float height;
+@property (nonatomic, assign) CGFloat width;
+@property (nonatomic, assign) CGFloat height;
 @property (nonatomic, readonly) CGRect bounds;
 @property (nonatomic, readonly) NSUInteger indexOfActiveLayer;
 @property (nonatomic, assign) BOOL snapToEdges;
